refactor: API Controller 통합
