+++ tamago-gateway/src/main/kotlin/tamago/server/gateway/presentation/user/v1/controller/UserController.kt
@@ -1,5 +1,7 @@
 package tamago.server.gateway.presentation.user.v1.controller
 
+import io.swagger.v3.oas.annotations.Operation
+import io.swagger.v3.oas.annotations.tags.Tag
 import jakarta.validation.Valid
 import org.springframework.security.access.prepost.PreAuthorize
 import org.springframework.web.bind.annotation.GetMapping
@@ -8,30 +10,32 @@ import org.springframework.web.bind.annotation.RequestBody
 import org.springframework.web.bind.annotation.RestController
 import tamago.server.core.user.UserCommandUseCase
 import tamago.server.core.user.domain.aggregate.User
-import tamago.server.gateway.presentation.user.v1.api.UserApi
+import tamago.server.gateway.common.annotation.CurrentUser
+import tamago.server.gateway.common.response.CustomResponse
 import tamago.server.gateway.presentation.user.v1.request.GoalKiloRequest
 import tamago.server.gateway.presentation.user.v1.request.NicknameRequest
 import tamago.server.gateway.presentation.user.v1.response.MeResponse
 import tamago.server.gateway.presentation.user.v1.response.RunningDataResponse
-import tamago.server.gateway.common.response.CustomResponse
-import tamago.server.gateway.common.annotation.CurrentUser
 import java.time.LocalDate
 import java.time.temporal.ChronoUnit
 
+@Tag(name = "User API", description = "유저 관련 API")
 @RestController
 class UserController(
     private val userCommandUseCase: UserCommandUseCase,
-) : UserApi {
+) {
 
+    @Operation(summary = "내 정보 조회", description = "현재 로그인한 유저의 정보를 반환합니다.")
     @PreAuthorize("hasRole('ROLE_USER')")
     @GetMapping("/api/v1/users/me")
-    override fun me(
+    fun me(
         @CurrentUser user: User
     ): CustomResponse<MeResponse> = CustomResponse.ok(MeResponse.from(user))
 
+    @Operation(summary = "애칭 설정하기", description = "타마고 애칭을 설정합니다.")
     @PreAuthorize("hasRole('ROLE_USER')")
     @PatchMapping("/api/v1/users/nickname")
-    override fun giveNickname(
+    fun giveNickname(
         @CurrentUser user: User,
         @Valid @RequestBody request: NicknameRequest
     ): CustomResponse<Void> {
@@ -39,9 +43,10 @@ class UserController(
         return CustomResponse.ok()
     }
 
+    @Operation(summary = "하루 러닝 목표 설정", description = "하루 러닝 목표를 정합니다.")
     @PreAuthorize("hasRole('ROLE_USER')")
     @PatchMapping("/api/v1/users/running-goal")
-    override fun setRunningGoal(
+    fun setRunningGoal(
         @CurrentUser user: User,
         @Valid @RequestBody request: GoalKiloRequest
     ): CustomResponse<Void> {
@@ -49,9 +54,10 @@ class UserController(
         return CustomResponse.ok()
     }
 
+    @Operation(summary = "누적 러닝 거리와 러닝 일수를 반환", description = "누적 러닝 거리와 러닝 일수를 반환합니다.")
     @PreAuthorize("hasRole('ROLE_USER')")
     @GetMapping("/api/v1/users/running-data")
-    override fun getRunningData(
+    fun getRunningData(
         @CurrentUser user: User
     ): CustomResponse<RunningDataResponse> {
         val runningDays = user.createdAt?.let {
